+++ src/lib.rs
@@ -4,3 +4,9 @@ mod utils;
 
 pub use cli::*;
 pub use process::*;
+
+#[warn(async_fn_in_trait)]
+#[enum_dispatch::enum_dispatch]
+pub trait CmdExecutor {
+    async fn execute(&self) -> anyhow::Result<()>;
+}