+++ src/process/mod.rs
@@ -2,6 +2,7 @@ mod b64;
 mod csv_convert;
 mod gen_pass;
 mod http_serve;
+mod jwt;
 mod text;
 
 pub use crate::utils::verify_input_file;
@@ -9,7 +10,8 @@ pub use b64::{decode_base64, encode_base64};
 pub use csv_convert::process_csv;
 pub use gen_pass::process_genpass;
 pub use http_serve::process_http_serve;
-pub use text::{generate_key, sign_text, verify_text};
+pub use jwt::{process_jwt_sign, process_jwt_verify};
+pub use text::{generate_key, process_decrypt, process_encrypt, sign_text, verify_text};
 
 #[cfg(test)]
 mod tests {